Add boolean and logic to nushell
This commit does a few things:

- Adds parsing for boolean and
- It allows us to handle arbitrary length boolean expressions
- It sets up a base upon which things like boolean or logic or other
  things like xor could use

The big thing here is that with arbitrary length boolean expressions we
can evaluate large complex expressions. The main motivation for this was
to allow the where command to have more expressivity. So instead of
something like:

ls | where size > 100b | where type == File

we can instead do the more readable expression of:

ls | where size > 100b && type == File

which more accurately represents the intent of the user.