Externals now spawn independently.
This commit changes the way we shell out externals when using the `$it` argument. Originally grabbing each row, making it a string, and constructing a single process call with `external row1 && external row2 && ..&& external rowN`. With this change we take a hit in performance since we spawn an external command process per row.

Also separated logic allowing us to prepare the values for feeding it whether it be through `$it` evaluation or as stdin.

The same code for spawning an external, wether it takes the row in the argument (`$it` replacement) or piped as stdin can be used (but the differences are visible since for now we spawn an external per row (`$it` case) as opposed to spawning one process and feeding all the rows into it's stdin.

This is deliberate so we can review it.

If we go in the direction of spawning an external and feeding it's stdin per row, the function `spawn` can be used without the `values` optional that is used to distinguish wether we want to feed into stdin or not. We have three new `dummies` for aiding in tests, in particular, the `chop` dummy.

`chop` does two distinct things:

1. chops the string passed as an argument (if given) and exits.
2. chops the input passed as stdin and exits.

With the current behavior if we do the following:

To clarify, let's assume we get from `ls | get name | first 2`:

```text
/home/andresrobalino> ls | get name | first 2
───┬─────────────┬
 # │ name        |
───┼─────────────┼
 0 │ andresA     |
 1 │ jonathanB   |
───┴─────────────┴
```

if we do (with the current behavior): `ls | get name | chop`

we get the output `andresA.jonathan`

if we choose spawning an external per row we'd get `andresjonathan` instead.