Move nu_errors::ParseError to nu_parser::errors::ParseError.
This is a first step to provide more information in the parse error with regards
to what successfully parsed, and the shape of what's incomplete. This will be
necessary information for the completion engine.

This does mean that other crates can't work with a `ParseError` directly, but no
crate really needs to at the moment. We can add `ParseError` back to `nu_errors`
if this was ever the case.