Make the sleep command respond to Ctrl+C
The internal sleep command is now implemented by a custom future.

The future spawns a thread that sleeps for the given duration, and
completes the future normally in that time.

The future also spawns a second thread that polls for Cltr+C presses.
If Ctrl+C occures, the future is finished early.